requirements.txt
python-telegram-bot==13.15
gspread==5.7.2
oauth2client==4.1.3